+++ fabric/src/main/java/com/evandev/reliable_replacer/compat/ModMenuIntegration.java
@@ -1,6 +1,6 @@
 package com.evandev.reliable_replacer.compat;
 
-import com.evandev.reliable_replacer.client.integration.ClothConfigIntegration;
+import com.evandev.reliable_replacer.client.integration.YACLIntegration;
 import com.evandev.reliable_replacer.platform.Services;
 import com.terraformersmc.modmenu.api.ConfigScreenFactory;
 import com.terraformersmc.modmenu.api.ModMenuApi;
@@ -8,8 +8,8 @@ import com.terraformersmc.modmenu.api.ModMenuApi;
 public class ModMenuIntegration implements ModMenuApi {
     @Override
     public ConfigScreenFactory<?> getModConfigScreenFactory() {
-        if (Services.PLATFORM.isModLoaded("cloth-config")) {
-            return ClothConfigIntegration::createScreen;
+        if (Services.PLATFORM.isModLoaded("yet_another_config_lib_v3")) {
+            return YACLIntegration::createScreen;
         }
         return null;
     }